fix warnings

+++ src/core/FileMasks.cpp
@@ -757,7 +757,7 @@ TCustomCommand::TCustomCommand()
 }
 //---------------------------------------------------------------------------
 void TCustomCommand::GetToken(
-  const std::wstring & Command, size_t Index, int &Len, char & PatternCmd)
+  const std::wstring & Command, size_t Index, size_t &Len, char & PatternCmd)
 {
   assert(Index < Command.size());
   const wchar_t * Ptr = Command.c_str() + Index - 1;
